+++ oereb_client/static/js/service/store.js
@@ -8,16 +8,35 @@ goog.require('oereb');
  * @ngInject
  * @ngdoc service
  * @ngName StoreService
+ *
+ * @param {angular.Scope} $rootScope The application's root scope.
+ * @param {string} oerebEventSymbolZoomEnabled Symbol zoom status event name.
+ * @param {string} oerebLocalStoragePrefix Prefix for the localStorage properties.
  */
-oereb.StoreService = function() {
-  if (!angular.isDefined(localStorage.blOerebHistory)) {
-    localStorage.blOerebHistory = angular.toJson([]);
+oereb.StoreService = function($rootScope, oerebEventSymbolZoomEnabled, oerebLocalStoragePrefix) {
+
+  this.$rootScope_ = $rootScope;
+  this.oerebEventSymbolZoomEnabled_ = oerebEventSymbolZoomEnabled;
+  this.historyProperty_ = oerebLocalStoragePrefix + 'OerebHistory';
+  this.availabilityProperty_ = oerebLocalStoragePrefix + 'OerebAvailability';
+  this.symbolZoomProperty_ = oerebLocalStoragePrefix + 'OerebSymbolZoom';
+
+  if (!angular.isDefined(localStorage[this.historyProperty_])) {
+    localStorage[this.historyProperty_] = angular.toJson([]);
   }
-  if (!angular.isDefined(localStorage.blOerebAvailability)) {
-    localStorage.blOerebAvailability = angular.toJson({
+
+  if (!angular.isDefined(localStorage[this.availabilityProperty_])) {
+    localStorage[this.availabilityProperty_] = angular.toJson({
       'show': true
     });
   }
+
+  if (!angular.isDefined(localStorage[this.symbolZoomProperty_])) {
+    localStorage[this.symbolZoomProperty_] = angular.toJson({
+      'show': true
+    });
+  }
+
 };
 
 /**
@@ -28,7 +47,7 @@ oereb.StoreService = function() {
  * @returns {Array} The array of previously loaded egrids.
  */
 oereb.StoreService.prototype.addRealEstate = function(realEstate) {
-  var history = angular.fromJson(localStorage.blOerebHistory);
+  var history = angular.fromJson(localStorage[this.historyProperty_]);
   var index = history.indexOf(angular.toJson(realEstate));
   if (index > -1) {
     history.splice(index, 1);
@@ -37,8 +56,8 @@ oereb.StoreService.prototype.addRealEstate = function(realEstate) {
     history.splice(0, 1);
   }
   history.push(angular.toJson(realEstate));
-  localStorage.blOerebHistory = angular.toJson(history);
-  var historyJsonContent = angular.fromJson(localStorage.blOerebHistory);
+  localStorage[this.historyProperty_] = angular.toJson(history);
+  var historyJsonContent = angular.fromJson(localStorage[this.historyProperty_]);
   history = [];
   angular.forEach(historyJsonContent, function (item) {
     history.push(angular.fromJson(item));
@@ -51,7 +70,7 @@ oereb.StoreService.prototype.addRealEstate = function(realEstate) {
  * @returns {Array} The array of previously loaded egrids.
  */
 oereb.StoreService.prototype.getHistory = function() {
-  var historyJsonContent = angular.fromJson(localStorage.blOerebHistory);
+  var historyJsonContent = angular.fromJson(localStorage[this.historyProperty_]);
   var history = [];
   angular.forEach(historyJsonContent, function (item) {
     history.push(angular.fromJson(item));
@@ -65,12 +84,27 @@ oereb.StoreService.prototype.getHistory = function() {
  * @returns {boolean} The availability layer visibility.
  */
 oereb.StoreService.prototype.showAvailability = function(show) {
-  var availability = angular.fromJson(localStorage.blOerebAvailability);
+  var availability = angular.fromJson(localStorage[this.availabilityProperty_]);
   if (angular.isDefined(show)) {
     availability['show'] = show;
-    localStorage.blOerebAvailability = angular.toJson(availability);
+    localStorage[this.availabilityProperty_] = angular.toJson(availability);
   }
   return availability['show'];
+};
+
+/**
+ * Enables/disables popovers for legend symbols and returns the current value.
+ * @param {boolean|undefined} show True or false to enable/disable popovers.
+ * @returns {boolean} True if popovers are enabled, false otherwise.
+ */
+oereb.StoreService.prototype.showSymbolZoom = function(show) {
+  var symbolZoom = angular.fromJson(localStorage[this.symbolZoomProperty_]);
+  if (angular.isDefined(show)) {
+    symbolZoom['show'] = show;
+    localStorage[this.symbolZoomProperty_] = angular.toJson(symbolZoom);
+    this.$rootScope_.$broadcast(this.oerebEventSymbolZoomEnabled_, show);
+  }
+  return symbolZoom['show'];
 };
 
 oereb.module.service('StoreService', oereb.StoreService);